requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.10.0
pydantic-settings==2.6.0

# Async Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Task Queue - Taskiq (replaces Celery)
taskiq[redis]==0.11.7
taskiq-redis==1.0.2
redis==5.0.8  # Compatible with taskiq-redis (>=5,<6)

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib==1.7.4

# HTTP & WebSocket
httpx==0.26.0
websockets==12.0
python-multipart==0.0.6

# AI Services
google-generativeai==0.4.0
instructor==1.7.0  # Pydantic-based LLM output validation

# Rate Limiting & Caching
slowapi==0.1.9
# fastapi-cache2[redis]==0.2.2  # Removed due to redis version conflict with taskiq-redis

# Document Processing
PyPDF2==3.0.1
olefile==0.47
beautifulsoup4==4.12.3

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
tenacity==9.0.0  # Updated for instructor compatibility
openpyxl==3.1.2
sendgrid==6.11.0

# Monitoring & Observability
prometheus-client==0.19.0
structlog==24.4.0
sentry-sdk[fastapi]==2.14.0

# Async Support
asgiref==3.7.2

# Development Tools
pre-commit==3.8.0
ruff==0.6.0
mypy==1.11.0

# Testing Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
aiosqlite==0.19.0
playwright==1.40.0

# ML Dependencies (used by prediction service)
joblib==1.3.2
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
